Completed Team color selection.

+++ src/com/becelot/spacerace/setup/ITeamBlockInteractionEvent.java
@@ -2,6 +2,9 @@ package com.becelot.spacerace.setup;
 
 import net.minecraft.entity.player.EntityPlayer;
 
+/*
+ * Interface for PlayerInteractEvent loclization
+ */
 public interface ITeamBlockInteractionEvent {
-	public void interact(EntityPlayer entityPlayer, int team, double x, double y, double z);
+	public void interact(EntityPlayer entityPlayer, int team, int x, int y, int z);
 }
